Moving routing_helpers.hpp and routing_helpers.cpp from map to storage.
