+++ map/map.pro
@@ -32,7 +32,6 @@ HEADERS += \
     mwm_url.hpp \
     place_page_info.hpp \
     reachable_by_taxi_checker.hpp \
-    routing_helpers.hpp \
     routing_manager.hpp \
     routing_mark.hpp \
     taxi_delegate.hpp \
@@ -67,7 +66,6 @@ SOURCES += \
     mwm_url.cpp \
     place_page_info.cpp \
     reachable_by_taxi_checker.cpp \
-    routing_helpers.cpp \
     routing_manager.cpp \
     routing_mark.cpp \
     taxi_delegate.cpp \
