+++ track_analyzing/track_analyzer/cmd_match.cpp
@@ -4,11 +4,10 @@
 #include "track_analyzing/track_matcher.hpp"
 #include "track_analyzing/utils.hpp"
 
-#include "map/routing_helpers.hpp"
-
 #include "routing_common/num_mwm_id.hpp"
 
 #include "storage/country_info_getter.hpp"
+#include "storage/routing_helpers.hpp"
 #include "storage/storage.hpp"
 
 #include "platform/platform.hpp"
